refactor: fix variable name typos and rename some class names
