+++ packages/services/src/identityProviders/IdentityProvider.ts
@@ -1,6 +1,9 @@
-import { IdentityProviderStrategy } from './identityProviderStrategy.js';
+export interface IdentityProviderStrategy {
+  getDlrUrl: (code: string, providerUrl: string) => Promise<string | null>;
+  getCode: (decodedText: string, formatName: string) => string;
+}
 
-export class IdentityProvder {
+export class IdentityProvider {
   private identityProviderStrategy: IdentityProviderStrategy;
   private identityProviderUrl: string;
