--- packages/services/src/identityProviders/IdentityProviderStrategy.ts
@@ -1,4 +1,0 @@
-export interface IdentityProviderStrategy {
-  getDlrUrl: (code: string, providerUrl: string) => Promise<string | null>;
-  getCode: (decodedText: string, formatName: string) => string;
-}
