+++ packages/services/src/identityProviders/index.ts
@@ -1,4 +1,3 @@
-export * from './gs1Provider.js';
-export * from './identityProviderStrategy.js';
-export * from './identityProvder.js';
+export * from './Gs1Provider.js';
+export * from './IdentityProvider.js';
 export * from './identityProviderServices.js';
